requirements.txt
gradio
fpdf2>=2.5.5
openai-agents
openai
pydantic
python-dotenv
requests
snscrape
duckduckgo-search
facebook-scraper
lxml_html_clean
